# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Vidhai ERP API — multi-site mushroom farming production orchestration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /batches:
    get:
      operationId: listBatches
      tags: [batches]
      summary: List all batches
      parameters:
        - name: locationId
          in: query
          required: false
          schema: { type: ["integer", "null"] }
        - name: stage
          in: query
          required: false
          schema: { type: ["string", "null"] }
        - name: status
          in: query
          required: false
          schema: { type: ["string", "null"] }
      responses:
        "200":
          description: Batches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Batch"
    post:
      operationId: createBatch
      tags: [batches]
      summary: Create a new batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}:
    get:
      operationId: getBatch
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Batch detail with full stage history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchDetail"
    patch:
      operationId: updateBatch
      tags: [batches]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
  /batches/{id}/advance:
    post:
      operationId: advanceBatchStage
      tags: [batches]
      summary: Advance batch to next stage
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StageAdvanceInput"
      responses:
        "200":
          description: Batch with updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Batch"
...
  /lab/batches:
    get:
      operationId: listLabBatches
      tags: [lab]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LabBatch" }
    post:
      operationId: createLabBatch
      tags: [lab]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabBatchInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabBatch" }
  /lab/batches/{id}:
    get:
      operationId: getLabBatch
      tags: [lab]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Lab batch detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabBatchDetail" }
    patch:
      operationId: updateLabBatch
      tags: [lab]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabBatchUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabBatch" }
  /lab/batches/{id}/advance:
    post:
      operationId: advanceLabBatchStage
      tags: [lab]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabStageAdvanceInput" }
      responses:
        "200":
          description: Stage advanced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabBatch" }
  /lab/batches/{id}/spawn-output:
    post:
      operationId: recordLabSpawnOutput
      tags: [lab]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabSpawnOutputInput" }
      responses:
        "201":
          description: Recorded
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabSpawnOutput" }
  /lab/spawn-transactions:
    get:
      operationId: listSpawnTransactions
      tags: [lab]
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SpawnTransaction" }
    post:
      operationId: createSpawnTransaction
      tags: [lab]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SpawnTransactionInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SpawnTransaction" }

  # ── OOTY ─────────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Batch:
      type: object
      required: [id, batchCode, locationId, locationCode, currentStage, status, createdAt, createdByName]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        locationId: { type: integer }
        locationCode: { type: string }
        currentStage:
          type: string
          enum: [PRE_WETTING, T1, T2, T3, T4, BULK_CHAMBER, QUALITY_CHECK, SPAWN_MIXING, DISPATCH, COMPLETED]
        status:
          type: string
          enum: [active, on_hold, completed, failed]
        nitrogenContent: { type: ["number", "null"] }
        targetBags: { type: ["integer", "null"] }
        actualBags: { type: ["integer", "null"] }
        spawnEntryId: { type: ["integer", "null"] }
        spawnBatchRef: { type: ["string", "null"] }
        spawnBatchType: { type: ["string", "null"] }
        dispatchLocationId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        createdByName: { type: string }
        stageEnteredAt: { type: ["string", "null"] }
        alertLevel: { type: ["string", "null"] }
    BatchDetail:
      type: object
      required: [id, batchCode, locationId, locationCode, currentStage, status, createdAt, createdByName, materials, stageLogs]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        locationId: { type: integer }
        locationCode: { type: string }
        currentStage:
          type: string
          enum: [PRE_WETTING, T1, T2, T3, T4, BULK_CHAMBER, QUALITY_CHECK, SPAWN_MIXING, DISPATCH, COMPLETED]
        status:
          type: string
          enum: [active, on_hold, completed, failed]
        nitrogenContent: { type: ["number", "null"] }
        targetBags: { type: ["integer", "null"] }
        actualBags: { type: ["integer", "null"] }
        spawnEntryId: { type: ["integer", "null"] }
        dispatchLocationId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        createdByName: { type: string }
        stageEnteredAt: { type: ["string", "null"] }
        alertLevel: { type: ["string", "null"] }
        materials:
          type: array
          items:
            $ref: "#/components/schemas/BatchMaterial"
        stageLogs:
          type: array
          items:
            $ref: "#/components/schemas/StageLog"
    BatchInput:
      type: object
      required: [locationId]
      properties:
        locationId: { type: integer }
        targetBags: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    BatchUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, on_hold, completed, failed]
        targetBags: { type: ["integer", "null"] }
        actualBags: { type: ["integer", "null"] }
        dispatchLocationId: { type: ["integer", "null"] }
        spawnEntryId: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
    StageAdvanceInput:
      type: object
      required: [nextStage]
      properties:
        nextStage:
          type: string
          enum: [PRE_WETTING, T1, T2, T3, T4, BULK_CHAMBER, QUALITY_CHECK, SPAWN_MIXING, DISPATCH, COMPLETED]
        notes: { type: ["string", "null"] }
        nh3Ppm: { type: ["number", "null"] }
        temperatureCelsius: { type: ["number", "null"] }
        spawnBatchRef: { type: ["string", "null"] }
        spawnBatchType:
          type: string
          enum: [internal, external]
          nullable: true
    BatchMaterial:
      type: object
      required: [id, batchId, materialId, materialName, unit, wetWeightKg, moisturePercent, nitrogenPercent]
      properties:
        id: { type: integer }
        batchId: { type: integer }
        materialId: { type: integer }
        materialName: { type: string }
        unit: { type: string }
        wetWeightKg: { type: number }
        moisturePercent: { type: number }
        nitrogenPercent: { type: number }
        dryWeightKg: { type: number }
        n2Kg: { type: number }
...
    StageLog:
      type: object
      required: [id, batchId, stage, enteredAt, enteredByName]
      properties:
        id: { type: integer }
        batchId: { type: integer }
        stage: { type: string }
        enteredAt: { type: string }
        exitedAt: { type: ["string", "null"] }
        enteredByName: { type: string }
        notes: { type: ["string", "null"] }
        nh3Ppm: { type: ["number", "null"] }
        temperatureCelsius: { type: ["number", "null"] }
...
    LabBatch:
      type: object
      required: [id, batchCode, locationId, locationCode, currentStage, status, createdAt]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        locationId: { type: integer }
        locationCode: { type: string }
        currentStage:
          type: string
          enum: [MEDIA_PREP, MOTHER_CULTURE, MILLET_1, MILLET_2, MOISTURE, AUTOCLAVE, INOCULATION, SHAKING_1, SHAKING_2, COMPLETED]
        status: { type: string }
        notes: { type: ["string", "null"] }
        alertLevel: { type: ["string", "null"] }
        createdAt: { type: string }
        stageEnteredAt: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
    LabBatchDetail:
      type: object
      required: [id, batchCode, locationId, locationCode, currentStage, status, createdAt, stageLogs, spawnOutputs]
      properties:
        id: { type: integer }
        batchCode: { type: string }
        locationId: { type: integer }
        locationCode: { type: string }
        currentStage: { type: string }
        status: { type: string }
        notes: { type: ["string", "null"] }
        alertLevel: { type: ["string", "null"] }
        createdAt: { type: string }
        stageEnteredAt: { type: ["string", "null"] }
        createdByName: { type: ["string", "null"] }
        stageLogs:
          type: array
          items: { $ref: "#/components/schemas/StageLog" }
        spawnOutputs:
          type: array
          items: { $ref: "#/components/schemas/LabSpawnOutput" }
    LabBatchInput:
      type: object
      properties:
        notes: { type: ["string", "null"] }
    LabBatchUpdate:
      type: object
      properties:
        status: { type: string }
        notes: { type: ["string", "null"] }
    LabStageAdvanceInput:
      type: object
      required: [nextStage]
      properties:
        nextStage:
          type: string
          enum: [MOTHER_CULTURE, MILLET_1, MILLET_2, MOISTURE, AUTOCLAVE, INOCULATION, SHAKING_1, SHAKING_2, COMPLETED]
        notes: { type: ["string", "null"] }
    LabSpawnOutput:
      type: object
      required: [id, batchId, strainName, quantityKg, status, createdAt]
      properties:
        id: { type: integer }
        batchId: { type: integer }
        strainName: { type: string }
        quantityKg: { type: number }
        producedAt: { type: ["string", "null"] }
        status:
          type: string
          enum: [available, consumed, allocated]
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
    LabSpawnOutputInput:
      type: object
      required: [strainName, quantityKg]
      properties:
        strainName: { type: string }
        quantityKg: { type: number }
        producedAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
    SpawnTransaction:
      type: object
      required: [id, transactionType, strainName, quantityKg, transactionDate, createdAt]
      properties:
        id: { type: integer }
        transactionType:
          type: string
          enum: [purchase, sale, internal_transfer]
        strainName: { type: string }
        quantityKg: { type: number }
        counterparty: { type: ["string", "null"] }
        unitPrice: { type: ["number", "null"] }
        transactionDate: { type: string }
        notes: { type: ["string", "null"] }
        labSpawnOutputId: { type: ["integer", "null"] }
        createdAt: { type: string }
    SpawnTransactionInput:
      type: object
      required: [transactionType, strainName, quantityKg, transactionDate]
      properties:
        transactionType:
          type: string
          enum: [purchase, sale, internal_transfer]
        strainName: { type: string }
        quantityKg: { type: number }
        counterparty: { type: ["string", "null"] }
        unitPrice: { type: ["number", "null"] }
        transactionDate: { type: string }
        notes: { type: ["string", "null"] }
        labSpawnOutputId: { type: ["integer", "null"] }

    # ── OOTY ─────────────────────────────────────────────────────────────────────
...